Move resolve link outside markdown extension core
