+++ src/Microsoft.DocAsCode.MarkdigEngine.Extensions/Inclusion/InclusionExtension.cs
@@ -5,11 +5,8 @@ namespace Microsoft.DocAsCode.MarkdigEngine.Extensions
 {
     using System.Threading;
     using Markdig;
-    using Markdig.Parsers;
     using Markdig.Parsers.Inlines;
     using Markdig.Renderers;
-    using Markdig.Syntax;
-    using Markdig.Syntax.Inlines;
 
     /// <summary>
     /// Extension to enable extension IncludeFile.
@@ -28,8 +25,6 @@ namespace Microsoft.DocAsCode.MarkdigEngine.Extensions
         {
             pipeline.BlockParsers.AddIfNotAlready<InclusionBlockParser>();
             pipeline.InlineParsers.InsertBefore<LinkInlineParser>(new InclusionInlineParser());
-
-            pipeline.DocumentProcessed += InclusionExtension.GetProcessDocumentDelegate(_context);
         }
 
         public void Setup(MarkdownPipeline pipeline, IMarkdownRenderer renderer)
@@ -46,46 +41,6 @@ namespace Microsoft.DocAsCode.MarkdigEngine.Extensions
                 if (!htmlRenderer.ObjectRenderers.Contains<HtmlInclusionBlockRenderer>())
                 {
                     htmlRenderer.ObjectRenderers.Insert(0, new HtmlInclusionBlockRenderer(_context, pipeline));
-                }
-            }
-        }
-
-        public static ProcessDocumentDelegate GetProcessDocumentDelegate(MarkdownContext context)
-        {
-            return document => UpdateLinks(document, context);
-        }
-
-        private static void UpdateLinks(MarkdownObject markdownObject, MarkdownContext context)
-        {
-            if (markdownObject == null || context == null) return;
-
-            if (markdownObject is ContainerBlock containerBlock)
-            {
-                foreach (var subBlock in containerBlock)
-                {
-                    UpdateLinks(subBlock, context);
-                }
-            }
-            else if (markdownObject is LeafBlock leafBlock)
-            {
-                if (leafBlock.Inline != null)
-                {
-                    foreach (var subInline in leafBlock.Inline)
-                    {
-                        UpdateLinks(subInline, context);
-                    }
-                }
-            }
-            else if (markdownObject is ContainerInline containerInline)
-            {
-                foreach (var subInline in containerInline)
-                {
-                    UpdateLinks(subInline, context);
-                }
-
-                if (markdownObject is LinkInline linkInline && !linkInline.IsAutoLink)
-                {
-                    linkInline.GetDynamicUrl = () => context.GetLink(linkInline.Url, InclusionContext.File, InclusionContext.RootFile, linkInline);
                 }
             }
         }
