+++ src/Microsoft.DocAsCode.MarkdigEngine/MarkdigMarkdownService.cs
@@ -34,8 +34,7 @@ namespace Microsoft.DocAsCode.MarkdigEngine
                 key => _parameters.Tokens.TryGetValue(key, out var value) ? value : null,
                 (code, message, origin, line) => Logger.LogWarning(message, null, InclusionContext.File.ToString(), line.ToString(), code),
                 (code, message, origin, line) => Logger.LogError(message, null, InclusionContext.File.ToString(), line.ToString(), code),
-                ReadFile,
-                GetLink);
+                ReadFile);
         }
 
         public MarkupResult Markup(string content, string filePath)
@@ -143,6 +142,7 @@ namespace Microsoft.DocAsCode.MarkdigEngine
 
             builder.UseDocfxExtensions(_context);
             builder.Extensions.Insert(0, new YamlHeaderExtension(_context));
+            builder.Extensions.Add(new ResolveLinkExtension());
 
             if (enableSourceInfo)
             {
@@ -160,15 +160,6 @@ namespace Microsoft.DocAsCode.MarkdigEngine
             }
 
             return builder.Build();
-        }
-
-        private static string GetLink(string path, object relativeTo, object resultRelativeTo, MarkdownObject origin)
-        {
-            if (InclusionContext.IsInclude && RelativePath.IsRelativePath(path) && PathUtility.IsRelativePath(path) && !RelativePath.IsPathFromWorkingFolder(path) && !path.StartsWith("#"))
-            {
-                return ((RelativePath)relativeTo + (RelativePath)path).GetPathFromWorkingFolder();
-            }
-            return path;
         }
 
         private (string content, object file) ReadFile(string path, object relativeTo, MarkdownObject origin)
